Fixed the duplicate symbol error. Got rid of warnings.
The symbol was duplicated because nExecutables declared in the chell.h header. So if the header was declared more than once, the variable would be defined twice.
The solution was to move that variable to the .c file and mark it as static. The static keyword makes sure that the symbol is local to that translation unit only, so even if there is a variable with the same name in other translation units, the names won't collide.

The warnings were simple enough to get rid of, I think the diff will show what I did.